03-device-simulators/requirements.txt
# Device Simulator Dependencies
paho-mqtt==1.6.1

